+++ package/lib/src/bar_chart.dart
@@ -34,6 +34,9 @@ class BarChart<D, T> extends StatefulWidget
   /// vertically, otherwise - horizontally.
   final Axis valueAxis;
 
+  /// The offset of outside labels from the edge of the bar.
+  final double labelsOffset;
+
   /// Whether to show zero bars.
   ///
   /// When the measure value is zero, the diagram doesn't show corresponding
@@ -181,6 +184,7 @@ class BarChart<D, T> extends StatefulWidget
     this.domainFormatter,
     this.measureFormatter,
     this.valueAxis = Axis.vertical,
+    this.labelsOffset = 2.0,
     this.showZeroValues = false,
     this.inverted = false,
 
@@ -303,6 +307,7 @@ class _BarChartState<D, T> extends State<BarChart<D, T>>
         animation: _currentAnimation,
         ticksResolver: _ticksResolver,
         measureFormatter: widget.measureFormatter,
+        labelsOffset: widget.labelsOffset,
         mainAxisTextStyle: widget.mainAxisTextStyle ?? TextStyle(
           fontSize: 12.0,
           color: colorScheme.onSurface,